refactor price and total to BigDecimal datatype

+++ src/main/java/com/github/MaxWilmes1/ObjectOrientation/RecapProject/SecondTry/Bestellung/Bestellung.java
@@ -2,11 +2,12 @@ package com.github.MaxWilmes1.ObjectOrientation.RecapProject.SecondTry.Bestellun
 
 import com.github.MaxWilmes1.ObjectOrientation.RecapProject.SecondTry.Produkt.Produkt;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record Bestellung(
         String id,
         List<Produkt> bestellteProdukte,
-        double total
+        BigDecimal total
 ) {
 }
